+++ src/wayparam/__init__.py
@@ -1,4 +1,4 @@
 # SPDX-License-Identifier: GPL-3.0
 
 __all__ = ["__version__"]
-__version__ = "0.9.0"
+__version__ = "1.0.0"